requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
jinja2>=3.1.0
weasyprint>=60.0
jsonschema>=4.19.0
requests>=2.31.0
pydantic>=2.5
pyyaml>=6.0
python-docx>=1.1
